[Gold V] Title: DKSH를 찾아라, Time: 0 ms, Memory: 2392 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/25427. DKSH를 찾아라/DKSH를 찾아라.cc
@@ -19,7 +19,7 @@ int main()
 	int N;
 	string str;
 	cin >> N >> str;
-	int d = 0, k = 0, s = 0, h = 0;
+	long long d = 0, k = 0, s = 0, h = 0;
 	for (int i = 1; i <= N; ++i) {
 		if (str[i - 1] == 'D') {
 			d++;
